Report errors to monitoring it they are handled in the inner express app (#707)
Currently, if you handle an error and send the response, but then call next(err), the error is not reported to monitoring
because response `finish` event is fired before the errorMetering middleware had the time to notice the error. We do the
reporting to monitoring in setImmediate to alleviate this.